+++ src/css/header.css
@@ -12,4 +12,11 @@
   background-color: lightgray;
 }
 
+.header fieldset ul{
+  display: flex;
+  align-items: center;
+  padding: var(--space-md);
+  background-color: lightgray;
+}
+
